Sort_modules function added to sort modules based on NetScore
Creation of sorting function that takes a channel of ratings (scores),
waits for it to be full and then sorts them using the built-in sorting
method for Slices, based on the NetScore in the rating. Outputs the
given ratings in a separate function. Testing is done in the main
function. rating struct changed slightly so that Url is a strig type
and License is of bool type.