Prototype for GraphQL API
GraphQL is used to receive number of watchers, total number of commits,
and number of people who star the repository in order to build the
correctness metric. Data is converted from string to json as the data
type specified as CorrectnessFactors which resembles the GraphQL query.

Should throw better errors.