Conversion of NPM url to Github url done (not well tested)
NPMtoGithubUrl takes in a npm url and converts it to a github url
through extraction of json data from npm registry and parsing of
strings until github url is found. Pretty inneficient and may be a
cause of lag because of huge json to string conversions and parsing.

Small fixes added to graphql.go to make it run.